uploaded_app/mock-testing-suite/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-generativeai==0.8.4
gspread==6.1.4
google-auth==2.37.0
selenium==4.27.1
webdriver-manager==4.0.2
